+++ Source/Public/Management/New/New-NinjaOneWindowsEventPolicyCondition.ps1
@@ -22,7 +22,7 @@ function New-NinjaOneWindowsEventPolicyCondition {
 	Param(
 		# The policy id to create the windows event policy condition for.
 		[Parameter(Mandatory, Position = 0, ValueFromPipeline, ValueFromPipelineByPropertyName)]
-		[Alias('id', 'policyId')]
+		[Alias('id')]
 		[Int]$policyId,
 		# An object containing the windows event policy condition to create.
 		[Parameter(Mandatory, Position = 1, ValueFromPipelineByPropertyName)]